KeyWorkProject/requirements.txt
reportlab
pdf2image
PyPDF2
pytesseract
Pillow
openai
python-dotenv
numpy